+++ Hashmaps/HashDoubleString.cpp
@@ -15,15 +15,15 @@
 class HashDoubleString {
 private:
     std::vector<std::pair<std::string,User>> table; /**< El vector que almacena los pares key-User. */
-    size_t table_size; /**< El tamaño de la tabla hash. >*/
+    unsigned long long table_size; /**< El tamaño de la tabla hash. >*/
 
 /**
      * @brief Primera funcion hash que crea un hash_value que empieza en 0 donde para cada caracter en la string key multiplica el hash_value por 33 y le suma el valor ascii del caracter, al resultado final le aplica el modulo con el tamaño de la tabla.
      * @param key La key a la que se le hará hash.
      * @return El valor hash.
      */
-    size_t hash1(const std::string& key) const {
-        size_t hash_value = 0;
+    unsigned long long hash1(const std::string& key) const {
+        unsigned long long hash_value = 0;
         for (char ch : key) {
             hash_value = hash_value * 33 + ch;
         }
@@ -35,8 +35,8 @@ private:
      * @param key La key a la que se le hará hash.
      * @return El valor hash.
      */
-    size_t hash2(const std::string& key) const {
-        size_t hash_value = 0;
+    unsigned long long hash2(const std::string& key) const {
+        unsigned long long hash_value = 0;
         for (char ch : key) {
             hash_value = hash_value * 17 + ch;
         }
@@ -48,7 +48,7 @@ public:
      * @brief Construye un objeto de HashDoubleString con el tamaño especificado.
      * @param size EL tamaño de la tabla hash.
      */
-    HashDoubleString(size_t size) : table_size(size) {
+    HashDoubleString(unsigned long long size) : table_size(size) {
         table.resize(size);
     }
 
@@ -58,8 +58,8 @@ public:
      * @param usuario El User a ser insertado.
      */
     void insert(std::string key, User usuario) {
-        size_t index = hash1(key);
-        size_t step = hash2(key);
+        unsigned long long index = hash1(key);
+        unsigned long long step = hash2(key);
         while (table[index].first != "" && table[index].first != "-1") {
             index = (index + step) % table.size();
         }
@@ -73,8 +73,8 @@ public:
      * @return True si la key es encontrada, False en otro caso.
      */
     bool search(std::string key, User& usuario) const {
-        size_t index = hash1(key);
-        size_t step = hash2(key);
+        unsigned long long index = hash1(key);
+        unsigned long long step = hash2(key);
         while (table[index].first != "" && table[index].first != key) {
             index = (index + step) % table.size();
         }
@@ -91,8 +91,8 @@ public:
      * @return True si la key es encontrada y eliminada, False en otro caso.
      */
     bool remove(std::string key) {
-        size_t index = hash1(key);
-        size_t step = hash2(key);
+        unsigned long long index = hash1(key);
+        unsigned long long step = hash2(key);
         while (table[index].first != "" && table[index].first != key) {
             index = (index + step) % table.size();
         }
