+++ Hashmaps/HashQuadInt.cpp
@@ -15,14 +15,14 @@
 class HashQuadInt {
 private:
     std::vector<std::pair<int,User>> table; /**< El vector que almacena los pares key-User. */
-    size_t table_size; /**< El tamaño de la tabla hash. >*/
+    unsigned long long table_size; /**< El tamaño de la tabla hash. >*/
 
     /**
      * @brief Funcion hash que retorna el modulo de la key con el tamaño de la tabla.
      * @param key La key a la que se le hará hash.
      * @return El valor hash.
      */
-    size_t hash(int key) const {
+    unsigned long long hash(int key) const {
         return key % table_size;
     }
 
@@ -31,7 +31,7 @@ public:
      * @brief Construye un objeto de HashQuadInt con el tamaño especificado.
      * @param size El tamaño de la tabla hash.
      */
-    HashQuadInt(size_t size) : table_size(size) {
+    HashQuadInt(unsigned long long size) : table_size(size) {
         table.resize(size);
     }
 
@@ -41,7 +41,7 @@ public:
      * @param usuario EL User a ser insertado.
      */
     void insert(int key, User usuario) {
-        size_t index = hash(key);
+        unsigned long long index = hash(key);
         int count = 0;
 
         while (table[index].first != 0 && table[index].first != -1) {
@@ -59,7 +59,7 @@ public:
      * @return True si la key es encontrada, False en otro caso.
      */
     bool search(int key, User& usuario) const {
-        size_t index = hash(key);
+        unsigned long long index = hash(key);
         int count = 0;
 
         while (table[index].first == 0 || table[index].first == -1) {
@@ -81,7 +81,7 @@ public:
      * @return True si la key es encontrada y eliminada, False en otro caso.
      */
     bool remove(int key) {
-        size_t index = hash(key);
+        unsigned long long index = hash(key);
         int count = 0;
 
         while (table[index].first == 0 || table[index].first == -1) {
